requirements.txt
moviepy==2.1.1
streamlit==1.40.2
edge_tts==6.1.19
fastapi==0.115.6
uvicorn==0.32.1
openai==1.56.1
faster-whisper==1.1.0
loguru==0.7.2
google-auth-oauthlib==1.0.0
google-api-python-client==2.92.0
google.generativeai==0.8.3
google-genai
dashscope==1.20.14
g4f==0.3.8.1
redis==5.2.0
python-multipart==0.0.19
streamlit-authenticator==0.4.1
pyyaml
